+++ core-bundle/src/Resources/contao/drivers/DC_Table.php
@@ -5570,19 +5570,19 @@ class DC_Table extends DataContainer implements \listable, \editable
 				// Sort by day
 				if (\in_array($GLOBALS['TL_DCA'][$this->strTable]['fields'][$field]['flag'], array(5, 6)))
 				{
-					$what = "UNIX_TIMESTAMP(FROM_UNIXTIME($what , '%%Y-%%m-%%d')) AS $what";
+					$what = "IF($what!='', UNIX_TIMESTAMP(FROM_UNIXTIME($what , '%%Y-%%m-%%d')), '') AS $what";
 				}
 
 				// Sort by month
 				elseif (\in_array($GLOBALS['TL_DCA'][$this->strTable]['fields'][$field]['flag'], array(7, 8)))
 				{
-					$what = "UNIX_TIMESTAMP(FROM_UNIXTIME($what , '%%Y-%%m-01')) AS $what";
+					$what = "IF($what!='', UNIX_TIMESTAMP(FROM_UNIXTIME($what , '%%Y-%%m-01')), '') AS $what";
 				}
 
 				// Sort by year
 				elseif (\in_array($GLOBALS['TL_DCA'][$this->strTable]['fields'][$field]['flag'], array(9, 10)))
 				{
-					$what = "UNIX_TIMESTAMP(FROM_UNIXTIME($what , '%%Y-01-01')) AS $what";
+					$what = "IF($what!='', UNIX_TIMESTAMP(FROM_UNIXTIME($what , '%%Y-01-01')), '') AS $what";
 				}
 			}
 
@@ -5614,7 +5614,7 @@ class DC_Table extends DataContainer implements \listable, \editable
 
 					foreach ($options as $k=>$v)
 					{
-						if ($v == '')
+						if ($v === '')
 						{
 							$options[$v] = '-';
 						}
@@ -5634,7 +5634,7 @@ class DC_Table extends DataContainer implements \listable, \editable
 
 					foreach ($options as $k=>$v)
 					{
-						if ($v == '')
+						if ($v === '')
 						{
 							$options[$v] = '-';
 						}
@@ -5660,7 +5660,7 @@ class DC_Table extends DataContainer implements \listable, \editable
 
 					foreach ($options as $k=>$v)
 					{
-						if ($v == '')
+						if ($v === '')
 						{
 							$options[$v] = '-';
 						}