+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change log
 
-## DEV
+## 4.4.36 (2019-03-25)
 
  * Make custom layout section titles and IDs mandatory (see #341).
  * Prevent using reserved layout section IDs in custom layout sections (see #301).